pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vision-ui"
version = "0.1.0"
description = "Screen-aware text & token budgeting utilities (UI_UX)"
readme = "UI_UX/README.md"
requires-python = ">=3.8"
license = { file = "LICENSE" }

authors = [
  { name = "shinychoes", email = "" }
]

keywords = [
  "screen",
  "ux",
  "ui",
  "budget",
  "summarization",
  "llm",
]

classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "License :: OSI Approved :: Apache Software License",
  "Intended Audience :: Developers",
  "Topic :: Software Development :: User Interfaces",
]

dependencies = [
  "rich>=14.0.0",
  "pytesseract>=0.3.0",
  "Pillow>=9.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0,<9.0.0",
  "pip-audit>=2.7.0,<3.0.0",
]
token = [
  "transformers>=4.40.0,<5.0.0",
]

[project.scripts]
vision-ui = "vision_ui.cli:main"

[tool.setuptools]
packages = ["UI_UX", "vision_ui"]

[tool.pytest.ini_options]
testpaths = ["tests"]
